# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Yang's Fried Rice - System API",
    "version": "1.0.0",
    "description": "Interactive Swagger/OpenAPI documentation for Menu Management, Administration, Order Oversight, and Reporting. Use 'Try it out' to execute live requests."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/reports/sales": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get daily sales metrics",
        "description": "Returns daily total revenue and order counts for the date range. Defaults to today. Requires admin or manager.",
        "operationId": "getDailySales",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD). Defaults to today."
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD). Defaults to today."
          }
        ],
        "responses": {
          "200": {
            "description": "Daily sales data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DailySalesRow"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/summary": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get platform performance summary",
        "description": "Returns aggregate revenue, total orders, and average order value. Defaults to today. Requires admin or manager.",
        "operationId": "getPlatformSummary",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Platform summary data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlatformSummary"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/preview": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get sales report preview",
        "description": "Combines summary metrics and daily breakdown for UI preview prior to PDF download. Requires admin or manager.",
        "operationId": "getReportPreview",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Report preview data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReportPreviewResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/pdf": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Generate PDF sales report",
        "description": "Generates formatted PDF report. Returns JSON with base64 data URI by default, or application/pdf binary file when ?format=raw. Requires admin or manager.",
        "operationId": "generateReportPDF",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date (YYYY-MM-DD)"
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "raw"
              ],
              "default": "json"
            },
            "description": "Output format: 'json' (datauri base64) or 'raw' (binary PDF download)"
          }
        ],
        "responses": {
          "200": {
            "description": "Generated PDF report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "pdf": {
                          "type": "string",
                          "description": "Base64 data URI string"
                        },
                        "format": {
                          "type": "string",
                          "example": "datauristring"
                        }
                      }
                    }
                  }
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports": {
      "post": {
        "tags": [
          "Reports"
        ],
        "summary": "Save report snapshot",
        "description": "Persists a report record into the database reports table. Requires admin or manager.",
        "operationId": "saveReport",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveReportInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report snapshot saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Report saved successfully"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Internal server error"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "DailySalesRow": {
        "type": "object",
        "required": [
          "date",
          "totalRevenue",
          "totalOrders"
        ],
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-09-10"
          },
          "totalRevenue": {
            "type": "number",
            "example": 8450.5
          },
          "totalOrders": {
            "type": "integer",
            "example": 24
          }
        }
      },
      "PlatformSummary": {
        "type": "object",
        "required": [
          "totalRevenue",
          "totalOrders",
          "averageOrderValue"
        ],
        "properties": {
          "totalRevenue": {
            "type": "number",
            "example": 42350
          },
          "totalOrders": {
            "type": "integer",
            "example": 120
          },
          "averageOrderValue": {
            "type": "number",
            "example": 352.92
          }
        }
      },
      "ReportPreviewResponse": {
        "type": "object",
        "properties": {
          "dateRange": {
            "type": "object",
            "properties": {
              "start_date": {
                "type": "string",
                "example": "2026-09-01"
              },
              "end_date": {
                "type": "string",
                "example": "2026-09-10"
              }
            }
          },
          "summary": {
            "$ref": "#/components/schemas/PlatformSummary"
          },
          "dailyBreakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailySalesRow"
            }
          }
        }
      },
      "SaveReportInput": {
        "type": "object",
        "properties": {
          "start_date": {
            "type": "string",
            "example": "2026-09-01"
          },
          "end_date": {
            "type": "string",
            "example": "2026-09-10"
          }
        }
      },
...
    }
  }
}